Update README Add TensorFlow library Add AI App for tests Add Representation App Add Data Generation App Fix Data generation process Add Async Logs Add Async storage

+++ src/main/java/com/cegesoft/data/metadata/FileMetadata.java
@@ -3,5 +3,5 @@ package com.cegesoft.data.metadata;
 import com.cegesoft.data.ByteStorable;
 
 public interface FileMetadata extends ByteStorable {
-    int getDataGroupSize();
+    short getDataGroupSize();
 }
